feat : refactor content script and improve inline button insertion
